refactor(analysis): extract JavaScript complexity span owner
