+++ crates/tokmd-analysis/src/complexity/details.rs
@@ -4,9 +4,11 @@ use tokmd_analysis_types::FunctionComplexityDetail;
 
 use super::map_language_for_complexity;
 
+mod javascript;
 mod python;
 mod rust;
 
+pub(super) use javascript::detect_fn_spans_js;
 pub(super) use python::detect_fn_spans_python;
 pub(super) use rust::detect_fn_spans_rust;
 
@@ -60,32 +62,6 @@ pub(super) fn extract_function_details(lang: &str, text: &str) -> Vec<FunctionCo
             }
         })
         .collect()
-}
-
-fn detect_fn_spans_js(lines: &[&str]) -> Vec<(usize, usize, String)> {
-    let mut spans = Vec::new();
-    let mut i = 0;
-    while i < lines.len() {
-        let trimmed = lines[i].trim();
-        let is_fn = trimmed.starts_with("function ")
-            || trimmed.starts_with("async function ")
-            || trimmed.starts_with("export function ")
-            || trimmed.starts_with("export async function ")
-            || trimmed.contains("=> {");
-        if is_fn && !trimmed.starts_with("//") {
-            let name = extract_js_fn_name(trimmed);
-            let start = i;
-            if let Some(end) = find_brace_end_at(lines, i) {
-                spans.push((start, end, name));
-                i = end + 1;
-            } else {
-                i += 1;
-            }
-        } else {
-            i += 1;
-        }
-    }
-    spans
 }
 
 fn detect_fn_spans_go(lines: &[&str]) -> Vec<(usize, usize, String)> {
@@ -155,34 +131,6 @@ fn find_brace_end_at(lines: &[&str], start_line: usize) -> Option<usize> {
         }
     }
     None
-}
-
-fn extract_js_fn_name(line: &str) -> String {
-    if let Some(idx) = line.find("function ") {
-        let after = &line[idx + 9..];
-        let name: String = after
-            .chars()
-            .take_while(|c| c.is_alphanumeric() || *c == '_' || *c == '$')
-            .collect();
-        if !name.is_empty() {
-            return name;
-        }
-    }
-    if let Some(paren_idx) = line.find('(') {
-        let before = line[..paren_idx].trim();
-        let name: String = before
-            .chars()
-            .rev()
-            .take_while(|c| c.is_alphanumeric() || *c == '_' || *c == '$')
-            .collect::<Vec<_>>()
-            .into_iter()
-            .rev()
-            .collect();
-        if !name.is_empty() {
-            return name;
-        }
-    }
-    "<anonymous>".to_string()
 }
 
 fn extract_go_fn_name(line: &str) -> String {
